requirements.txt
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
beautifulsoup4>=4.12.0
